Добавление фунцкионала

+++ app/src/main/java/com/example/backpackpicker/model/Item.java
@@ -2,10 +2,10 @@ package com.example.backpackpicker.model;
 
 public class Item {
     public final String name;
-    public final int weight;
-    public final int volume;
+    public final float weight;
+    public final float volume;
 
-    public Item(String name, int weight, int volume) {
+    public Item(String name, float weight, float volume) {
         this.name = name;
         this.weight = weight;
         this.volume = volume;
